Ajustes y mejoras Riesgos - JM
29-05-2026

+++ App.Infraestructure/IRepositories/IDocumentosControladosRepository.cs
@@ -14,6 +14,6 @@ namespace App.Infraestructure.IRepositories
         Task<TBL_doc_DocumentosModels> ObjDocumentosC(int DocumentoId);
         Task<List<TBL_doc_DocumentosModels>> DocumentosList(int EmpresaId);
         Task<PagedResult<ResponseTBL_doc_DocumentosModels>> GetListadoDocumentosFiltrosVista(int IdArea, int proceso_doc, string CodigoDoc, int IdProducto, int IdTipo, string Estado, int NivelSeguridad, 
-            int EmpresaId, int InIdSistema, int Usuario, string NombreDoc, int userId, int ElaboradoPor, int RevisadoPor, int AprobadoPor, string EstadoProceso, int Pagina, int pageSize, int codigoCargo);
+            int EmpresaId, int InIdSistema, long Usuario, string NombreDoc, long userId, int ElaboradoPor, int RevisadoPor, int AprobadoPor, string EstadoProceso, int Pagina, int pageSize, int codigoCargo);
     }
 }
